refactor: move date related utils to date.ts file
