+++ src/utils/date.ts
@@ -12,3 +12,18 @@ export const formatDate = (date: string | Date, stringFormat?: string) => {
   const localDate = dayjs(date).local();
   return stringFormat ? localDate.format(stringFormat) : localDate;
 };
+
+export const relativeDateTime = (ts: string | Date) => {
+  const t = dayjs.utc(ts).local();
+  const n = dayjs();
+
+  if (n.isSame(t, "day")) {
+    return t.fromNow();
+  }
+
+  if (n.diff(t, "days") === 1) {
+    return `${t.format("h:mm A")} yesterday`;
+  }
+
+  return t.format("h:mm A, dddd, MMMM D, YYYY");
+};
